Move LocalFeedLoader and FeedStore to production
